Fix issue due to upstream bug
The ReactOS LdrUnlockLoaderLock function was bugged. I submitted a PR to
ReactOS so nobody else faces this issue. It's worth noting that some
sites like geoffchappell.com still include the bug but there's no way I
can submit a fix for that.

Later, I may bring research back regarding the algorithm MS is using to
calculate the cookie value on a modern loader.